Edit project structure, lint files
