+++ .gitignore
@@ -1,0 +1,7 @@
+*.pyc
+bin
+include
+lib
+local
+pip-selfcheck.json
+share
